Struct serde_tagged::de::BoxFnSeed

pub struct BoxFnSeed<V>(_);

A boxed (immutable) closure that can be used as DeserializeSeed.

It additionally requires the wrapped closure to implement Sync which allows for easy static type-registry creation, e.g. in combination with BTreeMap<&'static str, _>.

Methods

impl<V> BoxFnSeed<V>

pub fn new<F>(func: F) -> Self where
F: FnSeed<V> + Sync + 'static,

Creates a new boxed closure from the given closure.