Trait serde_tagged::de::SeedFactory

pub trait SeedFactory<'de, T> {
    type Value;
    type Seed: DeserializeSeed<'de, Value = Self::Value>;
    fn seed<E>(self, tag: T) -> Result<Self::Seed, E>
    where
        E: Error;
}

A factory that creates a DeserializeSeed implementation based on a given tag.

Convenience implementations of this trait are provided for the standard map types with &'static str and String as key, as well as WithoutTag and WithTag.

Associated Types

type Value

The type of the value that will be produced by the DeserializeSeed implementation.

type Seed: DeserializeSeed<'de, Value = Self::Value>

The type of the DeserializeSeed implementation that will be returned by this factory.

Required Methods

fn seed<E>(self, tag: T) -> Result<Self::Seed, E> where
E: Error,

Returns the seed that should be used to deserialize a value tagged with the given tag.