Struct serde_tagged::ser::adj::tuple::Serializer

pub struct Serializer<'a, S, T: ?Sized + 'a> { /* fields omitted */ }

A serializer that serializes the specified tag and value as tuple.

The tag-value pair will be serialized as tuple where the first element will be the tag and the second element the tuple. The specified serializer performs the actual serialization and thus controls the data format. For more information on this tag-format, see the module documentation.

Warning

You should prefer the serialize function over this serializer implementation. To serialize a tuple, the serializer implementation may need to allocate memory on the heap. This can be avoided in the serialize function.

Methods

impl<'a, S, T: ?Sized> Serializer<'a, S, T> where
S: Serializer,
T: Serialize + 'a,

pub fn new(delegate: S, tag: &'a T) -> Self

Creates a new Serializer with the specified tag and underlying serializer.