Trait serde_tagged::ser::HasDelegate

pub trait HasDelegate {
    type Ok;
    type Error: Error;
    type Delegate: Serializer<Ok = Self::Ok, Error = Self::Error>;
    fn delegate(self) -> Self::Delegate;
}

A trait to provide access to a delegate serializer.

The delegate is expected to be responsible for the data-format of a serializer that implements this trait.

Associated Types

type Ok

type Error: Error

type Delegate: Serializer<Ok = Self::Ok, Error = Self::Error>

Required Methods

fn delegate(self) -> Self::Delegate

Returns the delegate-serializer of this object.